Clean ups, better exception management, checking if graph is loaded on 'cd'
