+++ src/turingdb/__init__.pyi
@@ -18,6 +18,7 @@ class TuringDB:
     def warmup(self, timeout: int = 5): ...
     def list_available_graphs(self) -> list[str]: ...
     def list_loaded_graphs(self) -> list[str]: ...
+    def is_graph_loaded(self) -> bool: ...
     def load_graph(self, graph_name: str, raise_if_loaded: bool = True): ...
     def create_graph(self, graph_name: str): ...
     def query(self, query: str): ...
